refactor : User, Dog 경로 변경 #7
